v2.0/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
pypdf
python-docx

langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-chroma
chromadb
langgraph

sqlalchemy
psycopg2-binary

httpx
rank-bm25
aiofiles
